x86: Pass legacy mmap base address as a parameter of arch_pick_mmap_base
This patch prepares for the merging of x86 mmap layout functions with
the generic ones. x86 is the only architecture that has the definition
of __TASK_UNMAPPED_BASE(task_size): in order to avoid duplicating this
macro in every other architecture, just remove its usage from
arch_pick_mmap_base that will become part of the generic functions.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>