x86: Use is_compat_task in arch_mmap_rnd
mmap_is_ia32(), in addition to testing if the current task is compat, also
checks if the system is 32bit. But HAVE_ARCH_MMAP_RND_COMPAT_BITS depends on
COMPAT, which in turn depends on IA32_EMULATION || x86_X32. And finally,
those configs depend on X86_64.
So simplify the test in arch_mmap_rnd so that it fits the generic version
defined with ARCH_WANT_DEFAULT_TOPDOWN_MMAP_LAYOUT.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>