x86: Use is_compat_task in arch_mmap_rnd
This commit replaces mmap_is_ia32() test with is_compat_task() in
arch_mmap_rnd(): it is possible to simplify this test because, in addition
to testing if the current task is compat, mmap_is_ia32() also checks if
the system is 32bit. But HAVE_ARCH_MMAP_RND_COMPAT_BITS depends on COMPAT,
which in turn depends on IA32_EMULATION || X86_X32. And finally, those
configs depend on X86_64.

So there is no need to check for X86_32 being defined and then we can
simplify arch_mmap_rnd to make it fit the generic version defined with
ARCH_WANT_DEFAULT_TOPDOWN_MMAP_LAYOUT.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>