mm: Add a define layer to allow more arch to fit mmap topdown functions
x86 does not compute mmap_base the same way that it is currently done in
the generic topdown mmap functions: that is because x86 allows 32bit task
to obtain 64bit mapping through 64bit syscall in compat mode or X86_X32.
Then the mmap_base always points to a 64bit base to handle those cases
and does not depend on task bitness as it is the case in STACK_TOP define.

In the same way, arch_mmap_rnd relies on the fact the task is compat or
not to compute the random gap for mmap. But for x86, this gap also does not
depend on task bitness.

So this commit introduces some defines for architectures to override to
handle such cases.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>