parisc: Use mmap_base, not mmap_legacy_base, as low_limit for bottomup mmap
Bottomup mmap scheme is used twice:

- for legacy mode, in which mmap_legacy_base and mmap_base are equal.

- in case of mmap failure in topdown mode, where there is no need to go
through the whole address space again for the bottomup fallback: the goal
of this fallback is to find, as a last resort, space between the topdown
mmap base and the stack, which is the only place not covered by the
topdown mmap.

Then this commit removes the usage of mmap_legacy_base field from parisc
code.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>