x86: Use mmap_*base, not mmap_*legacy_base, as low_limit for bottomup mmap
Bottomup mmap scheme is used twice:

- for legacy mode, in which mmap_legacy_base and mmap_compat_legacy_base are
respectively equal to mmap_base and mmap_compat_base.

- in case of mmap failure in topdown mode, where there is no need to go
through the whole address space again for the bottomup fallback: the goal
of this fallback is to find, as a last resort, space between the topdown
mmap base and the stack, which is the only place not covered by the
topdown mmap.

Then this commit removes the usage of mmap_legacy_base and
mmap_compat_legacy_base fields from x86 code.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>