riscv: Introduce CONFIG_RELOCATABLE
This config allows to compile kernel as PIE and to relocate it at runtime.
Compiling kernel as position independent code allows to relocate it at any
virtual address at runtime: this paves the way to KASLR and to 4-level page
table folding at runtime. Runtime relocation is possible since relocation
metadata are embedded into the kernel.

This is possible using the same compiler options used in arm64.

Note that:
- relocating at runtime introduces an overhead if even the kernel is
  loaded at the same address it was linked at.
- relocation granularity must be a multiple of 2MB as only PMD/PGD are
  allowed in early page table.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>