riscv: Introduce CONFIG_RELOCATABLE
This config allows to compile kernel as PIE and to relocate it at runtime.
Compiling kernel as position independent code allows to relocate it at any
virtual address at runtime: this paves the way to KASLR and to 4-level page
table folding at runtime. Runtime relocation is possible since relocation
metadata are embedded into the kernel.

Note that relocating at runtime introduces an overhead if even the kernel is
loaded at the same address it was linked at.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>