arch, scripts: Add script to check relocations at compile time
Relocating kernel at runtime is done very early in the boot process, so
it is not convenient to check for relocations there and react in case a
relocation was not expected.

Powerpc architecture has a script that allows to check at compile time
for such unexpected relocations: extract the common logic to scripts/
and add arch specific scripts triggered at postlink.

At the moment, powerpc, arm64 and riscv architectures take advantage of
this compile check.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>