clk: sunxi-ng: mux: Allow muxes to have keys
The muxes in the RTC can only be updated when setting a key field to a
specific value. Add a feature flag to denote muxes with this property.

Since so far the key value is always the same, it does not need to be
provided separately for each mux.

Series-changes: 3
 - Drop the SUNXI_CCU_MUX_HW_WITH_KEY macro, since it is no longer used.

Cover-changes: 3
 - Also drop the patch adding the SUNXI_CCU_MUX_DATA_WITH_GATE macro.

Signed-off-by: Samuel Holland <samuel@sholland.org>