riscv: Move KASAN mapping next to the kernel mapping
Now that KASAN_SHADOW_OFFSET is defined at compile time as a config,
this value must remain constant whatever the size of the virtual address
space, which is only possible by pushing this region at the end of the
address space next to the kernel mapping.

Also, take advantage of this modification to resize the linear mapping to
128GB instead of 124GB as it was since we have moved the kernel mapping
to the end of the address space.

Signed-off-by: Alexandre Ghiti <alexandre.ghiti@canonical.com>