ASoC: sun4i-spdif: Simplify code around optional resets
The driver does not need to care about which variants have a reset;
the devicetree binding already enforces that the necessary resources are
provided. Simplify the logic by always calling the optional getter,
which will return NULL if no reset reference is found.

Also clean up the error handling, which should not print a misleading
error in the EPROBE_DEFER case.

Signed-off-by: Samuel Holland <samuel@sholland.org>