UBUNTU: SAUCE: (no-up) Merge DMA non-coherency mechanisms
The BeagleV has devices that are not dma-coherent and use a mirror
mapping of the DRAM that is uncached to access those devices. The commit
924d6c2d34f2 ("RISC-V: Support non-coherent DMA operations") brings this
support for this board but then breaks other boards like the SiFive
Unmatched.

Allwinner takes advantage of the custom CMOs from T-HEAD core to guarantee
coherency in DMA transfers.

The SiFive Unmatched SoC is fully dma-coherent.

Make those 3 different ways to handle DMA coherency live together by
reading in the device tree very early how the dma coherency must be
handled, then all the DMA arch-specific callbacks use those information
to manage the coherency.

This commit also introduces new helpers to handle DIRECT_REMAP and
COHERENT_MMAP at runtime and not at compile-time as it is done for now
(Suggested-by: Atish Patra).

Fixes: 1c531d881adc ("riscv: cmo: Add dma-noncoherency support")
Fixes: 924d6c2d34f2 ("RISC-V: Support non-coherent DMA operations")
Signed-off-by: Alexandre Ghiti <alexandre.ghiti@canonical.com>