UBUNTU: SAUCE: (no-up) Fix the handling of absence of dma coherency
The BeagleV has devices that are not dma-coherent and use a mirror
mapping of the DRAM that is uncached to access those devices. The commit
924d6c2d34f2 ("RISC-V: Support non-coherent DMA operations") brings this
support for this board but then breaks other boards like the SiFive
Unmatched.

This patch moves the uncached offset to the device tree and any
cache operations regarding DMA transfers are executed only if such
offset exists.

This patch is not upstreamable as it fixes a vendor patch that won't get
upstreamed. The solution implemented here won't work when other boards
land but then we'll work on the more complex solution (and Atish waits
for such board to upstream its DMA patchset
https://patchwork.kernel.org/project/linux-riscv/list/?series=520541).

Tested on the Unmatched and the BeagleV boards.

Fixes: 924d6c2d34f2 ("RISC-V: Support non-coherent DMA operations")
Signed-off-by: Alexandre Ghiti <alexandre.ghiti@canonical.com>